Add another mode how to apply config changes
Currently, config changes are done by modifying the existing config.toml file.
Unfortunately, changes to the gitlab token itself and changes to the tags list
are not respected because the config.toml is not holding information about them.
Furthermore changing the config by modifying the config.toml is very time
consuming because the playbook has to run a lot of tasks.
For that reason, another way is introduced with this commit to apply config
changes by deregistering the gitlab runner and registering it again. This is
fairly fast and if no config changes are available, it has no runtime overhead
at all. Downside of this is, that the token is mandatory and saved in a hashed
form to a persistent file.