Allow unregistration of executors if not longer configured
When a runner executor is registered in gitlab but not longer configured
in ansible, currently the executor was untouched because it it not known
by this role. Now there is a new flag whcih controls if such executors should
be unregisterd, in case they are not longer configured in ansible.